+++ quill/src/bundled/fmt/os.cc
@@ -18,8 +18,8 @@
 #  include <sys/stat.h>
 #  include <sys/types.h>
 
-#  ifdef _WRS_KERNEL   // VxWorks7 kernel
-#    include <ioLib.h> // getpagesize
+#  ifdef _WRS_KERNEL    // VxWorks7 kernel
+#    include <ioLib.h>  // getpagesize
 #  endif
 
 #  ifndef _WIN32
@@ -182,10 +182,14 @@ void buffered_file::close() {
 }
 
 int buffered_file::descriptor() const {
-#ifdef fileno  // fileno is a macro on OpenBSD so we cannot use FMTQUILL_POSIX_CALL.
-  int fd = fileno(file_);
-#else
+#if !defined(fileno)
   int fd = FMTQUILL_POSIX_CALL(fileno(file_));
+#elif defined(FMTQUILL_HAS_SYSTEM)
+  // fileno is a macro on OpenBSD so we cannot use FMTQUILL_POSIX_CALL.
+#  define FMTQUILL_DISABLE_MACRO
+  int fd = FMTQUILL_SYSTEM(fileno FMTQUILL_DISABLE_MACRO(file_));
+#else
+  int fd = fileno(file_);
 #endif
   if (fd == -1)
     FMTQUILL_THROW(system_error(errno, FMTQUILL_STRING("cannot get file descriptor")));